requirements.txt
Flask==3.0.2
python-escpos==3.0a8
pyinstaller==6.4.0
